Streaming changes for HEVC cameras
