+++ lib/streaming/peer-connection.js
@@ -1,5 +1,5 @@
 // This code is largely copied from ring-client-api, but converted from Typescript
-// to straight Javascript and some code not required for ring-mqtt removed.
+// to native Javascript with custom logging for ring-mqtt and some unused code removed.
 // Much thanks to @dgreif for the original code which is the basis for this work.
 
 import { RTCPeerConnection, RTCRtpCodecParameters } from 'werift'
@@ -53,19 +53,6 @@ export class WeriftPeerConnection extends Subscribed {
                         ],
                         parameters: 'packetization-mode=1;profile-level-id=640029;level-asymmetry-allowed=1',
                     }),
-                    /*
-                    new RTCRtpCodecParameters({
-                        mimeType: 'video/H265',
-                        clockRate: 90000,
-                        rtcpFeedback: [
-                            { type: 'transport-cc' },
-                            { type: 'ccm', parameter: 'fir' },
-                            { type: 'nack' },
-                            { type: 'nack', parameter: 'pli' },
-                            { type: 'goog-remb' },
-                        ],
-                    }),
-                    */
                     new RTCRtpCodecParameters({
                         mimeType: "video/rtx",
                         clockRate: 90000,
