+++ lib/streaming/subscribed.js
@@ -1,5 +1,5 @@
 // This code is largely copied from ring-client-api, but converted from Typescript
-// to straight Javascript and some code not required for ring-mqtt removed.
+// to native Javascript with custom logging for ring-mqtt and some unused code removed.
 // Much thanks to @dgreif for the original code which is the basis for this work.
 
 export class Subscribed {
